Module serde_urlencoded::de

Deserialization support for the application/x-www-form-urlencoded format.

Structs

Deserializer

A deserializer for the application/x-www-form-urlencoded format.

Error

A minimal representation of all possible errors that can occur using the IntoDeserializer trait.

Functions

from_bytes

Deserializes a application/x-www-form-urlencoded value from a &[u8].

from_reader

Convenience function that reads all bytes from reader and deserializes them with from_bytes.

from_str

Deserializes a application/x-www-form-urlencoded value from a &str.